// cosmowcl/src/app/shared/models/api/cosmoecm/identitaUtente.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * acta.v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 */
export interface IdentitaUtente { 
    id: string;
    identificativoAOO?: string;
    codiceAOO?: string;
    descrizioneAOO?: string;
    identificativoNodo?: string;
    codiceNodo?: string;
    descrizioneNodo?: string;
    identificativoStruttura?: string;
    codiceStruttura?: string;
    descrizioneStruttura?: string;
}